requirements.txt
appdirs==1.4.4
appnope==0.1.2
backcall==0.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==4.2.2
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cssselect==1.1.0
debugpy==1.5.1
decorator==5.1.0
entrypoints==0.3
fake-useragent==0.1.11
idna==2.5
importlib-metadata==4.10.0
ipykernel==6.4.2
ipython==7.28.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
litereval==0.0.11
lxml==4.7.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.1
nltk==3.6.7
numpy==1.22.0
pandas==1.3.5
parse==1.19.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.21
psutil==5.9.0
ptyprocess==0.7.0
pyee==8.2.2
Pygments==2.10.0
pyppdf==0.1.2
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
regex==2021.11.10
requests==2.27.1
requests-html==0.10.0
ruby==0.0.2
schedule==1.1.0
six==1.16.0
soupsieve==2.3.1
telegram-send==0.25
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
w3lib==1.22.0
wcwidth==0.2.5
websockets==9.1
zipp==3.7.0
APScheduler==3.0.0
